Predictor/requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.2
joblib==1.3.2
pyyaml==6.0.1
xgboost==2.0.3
lightgbm==4.2.0